Make loading screen match launch screen
When the app is loaded into memory the subsequent transition of (
Launch Screen -> Loading Screen -> Main UI ) was visually unappealing.

This was mostly due to the fact that the loading screen is usually not
displayed for a significant amount of time (0-10s)

The solution was to make the loading screen match the launch screen
visually.