Fix error set size increasing indefinitely
Motivation:
All error events (even with the same error type) are added
to `errorAlerts` set.
This is unnenesary and uses more memory than we need.
Since we only need to store the last error event of the same
error type to know whether toshow an alert dialog or not.

Modifications:
Add a insert with replacement function if some path of the
 value to be inserted is the same.

Result:
Reduce memory usage by removing data no longer needed.