Fix crash where valid state handled as an error
Motivation:
App crashes during a state update in `PsiCashViewController`.
The reason that this crash is rarely seen is because the 
subscription to the signal is not called while purchasing.
Without `skipRepeats()` operator, this wouldn't be the case.

Modifications:
Add valid states as no-ops to the subscription function.

Result:
Fixes crash when signal subscription is called while purchasing.