Update VPN state with provider sync errors
Motivation:
Sync errors with the tunnel providers are only logged silently.
To be able to act on the errors (for example showing an alert
of what went wrong to the user), the error result should be added as
part of the app's state.

Modifications:
- Replace boolean `pendingProviderSync` flag with
  `Pending<ErrorEvent<TunnelProviderSyncedState.SyncError>?>` type.
- Update all previous usages of `pendingProviderSync`.

Result:
Sync errors are now part of the app's state.