Updated flow for tunnel provider with unknown sync state
Motivation:
Whenever a sync request is made to the tunnel provider, and the request
fails with an error, an active approach is taken to resolve the error,
instead of only logging it passively.

Previous implementation had a hard requirement of tunnel provider
having to respond within 5 seconds to sync request. However, in practice
this is found to be unnecessary and results in too many error alerts
being shown to the user.

Modifications:
- After a failed sync request, start or stop the VPN
  based on `NEVPNConnection.status` object's value.

- Present an alert to the user if there's a mismatch between
  tunnel intent and tunnel start only if tunnel intent is `.start`.
  Waits for 5 seconds for the mismatch to be resolved.

- Added `disableConnectOnDemand` Debug flag.