Fix: weak reference should be strong
The returned signal should hold a strong reference to FeedbackUpload so
it is not deallocated if no other reference is held. While this is not
an issue in the current way that FeedbackUpload is used, this could
result in an observer waiting indefinitely if FeedbackUpload is
deallocated before the completion handler from FeedbackHandler fires.