Fix displaying disallowed-traffic server alert for subscribing users
Motivation:
Previously disallowed-traffic alert generated by the extension was
conditioned only on the presence of an active speed-boost or
subscription authorization. This left out the case where the
user (potentially) has a valid subscription for which an authorization
has not been retrieved yet.

In this scenario it is acceptable to trust the receipt,
since the only downside (e.g. tampered system clock) is
an alert not getting displayed.

Modifications:
- Conditioned presenting disallowed-traffic alert on the local app
  receipt, and also on subscription or speed-boost authorizations.
- Added same condition to the alert presented by the container.

Result:
Fixes issue where disallowed-traffic alert is displayed while
an authorization is being retrieved for a recently bought subscription.