Added return type to ad controller present function
AdMob does not document the reasons an ad presentation can fail,
therefore we consider all of them fatal (as in the loaded ad can't
be used again, and another one should be loaded).

This commit adds a return type to the ad controller's present function,
and if presentation fails as determined by the function, it returns
an error message (as opposed to for example updating the status value).