Restricted state property-wrapper to the main thread
This fixes the following bug:
tunnelConnectionRefSignal is called on a background thread
which accesses AppState current value directly synchronously.
This can cause a memory exception.

If signalProducer is called from a thread other than the main-thread,
then then an async work is dispatched to the main-queue to read
AppState's current value.