Start VPN if last stop was reassertion timeout
NEProviderStopReason ConnectionFailed and IdleTimeout
indicate that the extension was stopped because it was in
the reasserting state for too long. After the extension is
stopped with either reason, sometimes "Connect On Demand"
rules bring the extension back up, which causes our code to
put it into zombie mode if the user does not have an active
authorization. Instead start the VPN.

Related https://forums.developer.apple.com/forums/thread/744949.